Rename  component folder
