+++ src/App.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import Calculator from './component/calculator';
+import Calculator from './components/calculator';
 import './App.css';
 
 class App extends React.Component {
